update webgl example
