+++ _PR.md
@@ -1,0 +1,23 @@
+# Changes
+
+## TODO
+
+- update all examples
+- update the documentation to mention trunk
+
+## Examples
+
+Fixes: #1368
+
+## 1. counter
+
+## 2. crm
+
+Modified to make the example more modular.
+What used to be a single component handling a lot of message spaghetti is now split into dedicated components.
+
+## 3. custom_components
+
+**Should be removed or updated**.
+
+## 26. webgl
