+++ examples/webgl/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2018"
 
 [dependencies]
 js-sys = "0.3"
-yew = { path = "../../yew" }
 wasm-bindgen = "0.2"
+yew = { path = "../../yew" }
 
 [dependencies.web-sys]
 version = "0.3"
