+++ examples/webgl/README.md
@@ -1,4 +1,13 @@
-## Yew WebGL Demo
+# WebGL Example
 
 This is a simple demo using WebGL with Yew to initialize the GL context, create
 a render loop, and draw to the canvas with basic shaders using `web-sys`.
+
+## Concepts
+
+- Accessing a DOM element using `NodeRef`.
+- Using Javascript APIs with `web-sys`.
+
+## Improvements
+
+- Use a much more flashy shader
