+++ examples/webgl/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Yew - WebGL</title>
+  </head>
+
+  <body></body>
+</html>
